+++ flow-server/src/test/java/com/vaadin/flow/server/frontend/NodeExecutorTest.java
@@ -1,0 +1,83 @@
+/*
+ * Copyright 2000-2019 Vaadin Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ *
+ */
+
+package com.vaadin.flow.server.frontend;
+
+import java.io.File;
+import java.lang.annotation.Annotation;
+import java.util.ArrayList;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
+
+import org.junit.Assert;
+import org.junit.Before;
+import org.junit.Test;
+import org.mockito.Mockito;
+import org.mockito.internal.matchers.Any;
+import org.mockito.invocation.InvocationOnMock;
+import org.mockito.stubbing.Answer;
+
+public class NodeExecutorTest extends NodeUpdateTestBase {
+
+    private File importsFile;
+    private File nodeModulesPath;
+
+    NodeExecutor node;
+
+    ClassPathIntrospector.ClassFinder classFinder;
+//    ClassPathIntrospector.ClassFinder classFinderMock;
+
+    @Before
+    public void setup() throws Exception {
+
+        File tmpRoot = temporaryFolder.getRoot();
+        importsFile = new File(tmpRoot, "flow-imports.js");
+        nodeModulesPath = new File(tmpRoot, "node_modules");
+
+//        classFinderMock = Mockito.mock(ClassPathIntrospector.ClassFinder.class);
+        classFinder = getClassFinder();
+
+//        Mockito.when(classFinderMock.getAnnotatedClasses(Mockito.any()))
+//                .thenAnswer(new Answer<Object>() {
+//                    public Object answer(InvocationOnMock invocation) {
+//                        return classFinder.getAnnotatedClasses(
+//                                (Class<? extends Annotation>) invocation
+//                                        .getArguments()[0]);
+//                    }
+//                });
+
+//        Mockito.when(classFinderMock.getAnnotatedClasses(Mockito.any()))
+//                .thenAnswer(new Answer<Object>() {
+//                    public Object answer(InvocationOnMock invocation) {
+//                        return classFinder.getAnnotatedClasses(
+//                                (Class<? extends Annotation>) invocation
+//                                        .getArguments()[0]);
+//                    }
+//                });
+
+        node = new NodeExecutor.Builder(classFinder, importsFile, tmpRoot,
+                nodeModulesPath, true).build();
+
+        createExpectedImports(importsFile.getParentFile(), nodeModulesPath);
+    }
+
+    @Test
+    public void should_ScanAnnotations_Once() {
+        node.execute();
+    }
+}
